Update Outstatic instructions for commit messaging
